pr-review-agent/requirements.txt
requests>=2.31.0,<3.0.0
python-docx>=1.1.2,<2.0.0
pytest>=8.0.0,<9.0.0
